refactor: Cleanup and simplify message module imports
- Remove unnecessary import of CertificateHashDataType in get_certificate_chain_status
- Simplify import paths for custom data and status types
- Update validation attributes for get_variables request and response
- Streamline import statements in log_status_notification and notify_der_start_stop
